Flux.train!: Call callback with info about previous training pass.
Call each callback passed to Flux.train! with a CallbackInfo struct.
Don't pass the CallbackInfo struct to callbacks that don't take any
arguments. Callbacks that are wrapped in `throttle` need to explicitly
ignore the callback struct, because `throttle` takes an arbitrary
number of arguments and Flux is (for now) not treating it in a special
way. Export CallbackInfo so that users can access its docstring.

Fixes #489.